Phonepe Project - GCP/requirements.txt
streamlit>=1.24.0
pandas>=2.0.3
plotly>=5.18.0
streamlit-option-menu>=0.5.1
google-cloud-storage>=2.10.0
google-auth>=2.22.0
requests>=2.31.0
